refactor(node-vfs): rename sandbox module to backend
