+++ libs/providers/node-vfs/src/sandbox.test.ts
@@ -1,7 +1,7 @@
 /* oxlint-disable no-instanceof/no-instanceof */
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 
-import { VfsSandbox } from "./sandbox.js";
+import { VfsSandbox } from "./backend.js";
 import { VfsSandboxError } from "./types.js";
 
 describe("VfsSandbox", () => {
